[backend] add tests
