+++ opencti-platform/opencti-graphql/tests/01-unit/utils/access-test.ts
@@ -1,13 +1,106 @@
 import { describe, it, expect } from 'vitest';
 import { v4 as uuid } from 'uuid';
-import { isMarkingAllowed, isOrganizationAllowed } from '../../../src/utils/access';
+import { ADMINISTRATOR_ROLE, checkUserCanAccessStixElement, isMarkingAllowed, isOrganizationAllowed, KNOWLEDGE_KNUPDATE_KNMANAGEAUTHMEMBERS } from '../../../src/utils/access';
 import type { BasicStoreCommon } from '../../../src/types/store';
 import { MARKING_TLP_AMBER, MARKING_TLP_CLEAR, MARKING_TLP_GREEN, MARKING_TLP_RED } from '../../../src/schema/identifier';
 import type { AuthUser } from '../../../src/types/user';
 import type { BasicStoreSettings } from '../../../src/types/settings';
-import { PLATFORM_ORGANIZATION, TEST_ORGANIZATION } from '../../utils/testQuery';
+import { PLATFORM_ORGANIZATION, TEST_ORGANIZATION, testContext } from '../../utils/testQuery';
 import { RELATION_GRANTED_TO } from '../../../src/schema/stixRefRelationship';
 import type { BasicStoreEntityOrganization } from '../../../src/modules/organization/organization-types';
+import type { StixObject } from '../../../src/types/stix-common';
+
+const report = {
+  id: 'report--f3e554eb-60f5-587c-9191-4f25e9ba9f32',
+  spec_version: '2.1',
+  type: 'report',
+  extensions: {
+    'extension-definition--ea279b3e-5c71-4632-ac08-831c66a786ba': {
+      extension_type: 'property-extension',
+      id: 'f13cd64f-9268-4d77-9850-eb6fbe322463',
+      type: 'Report',
+      creator_ids: [
+        '88ec0c6a-13ce-5e39-b486-354fe4a7084f'
+      ],
+      authorized_members: [
+        {
+          id: '88ec0c6a-13ce-5e39-b486-354fe4a7084f',
+          access_right: 'admin'
+        },
+        {
+          id: '55ec0c6a-13ce-5e39-b486-354fe4a7084f',
+          access_right: 'view'
+        },
+      ],
+    },
+  },
+} as unknown as StixObject;
+
+const user_is_allowed = {
+  administrated_organizations: [],
+  entity_type: 'User',
+  id: '55ec0c6a-13ce-5e39-b486-354fe4a7084f',
+  internal_id: '55ec0c6a-13ce-5e39-b486-354fe4a7084f',
+  individual_id: undefined,
+  organizations: [],
+  name: '',
+  user_email: '',
+  roles: [ADMINISTRATOR_ROLE],
+  groups: [],
+  capabilities: [{ name: KNOWLEDGE_KNUPDATE_KNMANAGEAUTHMEMBERS }],
+  all_marking: [],
+  inside_platform_organization: true,
+  allowed_marking: [],
+  default_marking: [],
+  origin: { referer: 'test', user_id: '55ec0c6a-13ce-5e39-b486-354fe4a7084f' },
+  api_token: 'd434ce02-e58e-4cac-8b4c-42bf16748e56',
+  account_status: '',
+  account_lock_after_date: undefined,
+  effective_confidence_level: {
+    max_confidence: 100,
+    overrides: [],
+  },
+  user_confidence_level: {
+    max_confidence: 100,
+    overrides: [],
+  },
+  max_shareable_marking: [],
+  restrict_delete: false,
+  no_creators: false,
+};
+
+const user_is_not_allowed = {
+  administrated_organizations: [],
+  entity_type: 'User',
+  id: '48ec0c6a-13ce-5e39-b486-354fe4a7084f',
+  internal_id: '48ec0c6a-13ce-5e39-b486-354fe4a7084f',
+  individual_id: undefined,
+  organizations: [],
+  name: '',
+  user_email: '',
+  roles: [ADMINISTRATOR_ROLE],
+  groups: [],
+  capabilities: [{ name: KNOWLEDGE_KNUPDATE_KNMANAGEAUTHMEMBERS, }],
+  all_marking: [],
+  inside_platform_organization: false,
+  allowed_marking: [],
+  default_marking: [],
+  origin: { referer: 'test', user_id: '48ec0c6a-13ce-5e39-b486-354fe4a7084f' },
+  api_token: 'd434ce02-e58e-4cac-8b4c-42bf16748e48',
+  account_status: '',
+  account_lock_after_date: undefined,
+  effective_confidence_level: {
+    max_confidence: 100,
+    overrides: [],
+  },
+  user_confidence_level: {
+    max_confidence: 100,
+    overrides: [],
+  },
+  max_shareable_marking: [],
+  restrict_delete: false,
+  no_creators: false,
+};
 
 describe('Check markings test coverage', () => {
   it('should element with no marking be allowed', async () => {
@@ -66,7 +159,7 @@ describe('Check organization access for element.', () => {
       platform_organization: undefined,
     };
 
-    expect(isOrganizationAllowed(element as BasicStoreCommon, user as AuthUser, settings as BasicStoreSettings)).toBeTruthy();
+    expect(isOrganizationAllowed(element as BasicStoreCommon, user as AuthUser, settings.platform_organization as string)).toBeTruthy();
   });
 
   it('should element not shared be allowed to user in platform organization', async () => {
@@ -90,7 +183,7 @@ describe('Check organization access for element.', () => {
       platform_organization: PLATFORM_ORGANIZATION.name,
     };
 
-    expect(isOrganizationAllowed(element as BasicStoreCommon, user as AuthUser, settings as BasicStoreSettings)).toBeTruthy();
+    expect(isOrganizationAllowed(element as BasicStoreCommon, user as AuthUser, settings.platform_organization as string)).toBeTruthy();
   });
 
   it('should element not shared not be allowed to user in another organization', async () => {
@@ -114,7 +207,7 @@ describe('Check organization access for element.', () => {
       platform_organization: PLATFORM_ORGANIZATION.name,
     };
 
-    expect(isOrganizationAllowed(element as BasicStoreCommon, user as AuthUser, settings as BasicStoreSettings)).toBeFalsy();
+    expect(isOrganizationAllowed(element as BasicStoreCommon, user as AuthUser, settings.platform_organization as string)).toBeFalsy();
   });
 
   it('should element shared to user organization be allowed', async () => {
@@ -140,6 +233,17 @@ describe('Check organization access for element.', () => {
       platform_organization: PLATFORM_ORGANIZATION.name,
     };
 
-    expect(isOrganizationAllowed(element as BasicStoreCommon, user as AuthUser, settings as BasicStoreSettings)).toBeTruthy();
+    expect(isOrganizationAllowed(element as BasicStoreCommon, user as AuthUser, settings.platform_organization as string)).toBeTruthy();
+  });
+});
+
+describe('checkUserCanAccessStixElement testing', async () => {
+  it('user in auth members should access element', async () => {
+    const hasAccess = await checkUserCanAccessStixElement(testContext, user_is_allowed, report, true);
+    expect(hasAccess).toEqual(true);
+  });
+  it('user not in auth members should not access element', async () => {
+    const hasAccess = await checkUserCanAccessStixElement(testContext, user_is_not_allowed, report, true);
+    expect(hasAccess).toEqual(false);
   });
 });
